fix(daemon): split fd_limit by GOOS to satisfy unconvert lint
The uint64 cast at fd_limit_unix.go:24 was added in 63024de for
freebsd, where syscall.Rlimit.Cur is int64. On linux/darwin (where
Cur is already uint64) the cast is a no-op and unconvert flagged it,
breaking CI lint.

Split into two files with explicit build tags:
  - fd_limit_unix.go: !windows && !freebsd — returns rlim.Cur directly
  - fd_limit_freebsd.go: freebsd — keeps the uint64() cast it needs

Each platform sees only the variant the linter agrees with. Verified
darwin, linux, and freebsd cross-compile and `golangci-lint run -c
.config/golangci.yml ./internal/daemon/` reports 0 issues.

+++ internal/daemon/fd_limit_unix.go
@@ -1,4 +1,4 @@
-//go:build !windows
+//go:build !windows && !freebsd
 
 package daemon
 
@@ -11,15 +11,12 @@ import "syscall"
 //
 // Unix-only because syscall.Getrlimit / syscall.Rlimit / syscall.RLIMIT_NOFILE
 // are guarded by `//go:build unix` upstream and don't exist for GOOS=windows.
-// See fd_limit_windows.go for the Windows stub.
-//
-// rlim.Cur is uint64 on linux and darwin but int64 on freebsd; cast to keep
-// the cross-GOOS return type stable. Negative values can't occur for
-// RLIMIT_NOFILE in practice.
+// See fd_limit_windows.go for the Windows stub and fd_limit_freebsd.go for
+// the freebsd variant where rlim.Cur is signed and needs an explicit cast.
 func CurrentProcessFDLimit() uint64 {
 	var rlim syscall.Rlimit
 	if err := syscall.Getrlimit(syscall.RLIMIT_NOFILE, &rlim); err != nil {
 		return 0
 	}
-	return uint64(rlim.Cur)
+	return rlim.Cur
 }
